refactor: Better naming
