+++ e2e/fixtures/views/MetricsReducerView.ts
@@ -13,17 +13,17 @@ export class MetricsReducerView extends DrilldownView {
     return super.goto(new URLSearchParams([...this.urlParams, ...new URLSearchParams(urlSearchParams)]));
   }
 
-  /* Header controls */
+  /* List controls */
 
-  getHeaderControls() {
-    return this.getByTestId('header-controls');
+  getListControls() {
+    return this.getByTestId('list-controls');
   }
 
-  async assertHeaderControls() {
-    const headerControls = this.getHeaderControls();
+  async assertListControls() {
+    const listControls = this.getListControls();
 
-    await expect(headerControls.getByText('Group by label')).toBeVisible();
-    await expect(headerControls.getByText('Sort by')).toBeVisible();
+    await expect(listControls.getByText('Group by label')).toBeVisible();
+    await expect(listControls.getByText('Sort by')).toBeVisible();
 
     await expect(this.getQuickFilterInput()).toBeVisible();
     await expect(this.getLayoutSwitcher()).toBeVisible();
@@ -34,7 +34,7 @@ export class MetricsReducerView extends DrilldownView {
   /* Quick filter */
 
   getQuickFilterInput() {
-    return this.getHeaderControls().getByPlaceholder('Search metrics');
+    return this.getListControls().getByPlaceholder('Search metrics');
   }
 
   async assertQuickFilter(expectedValue: string, expectedResultsCount: number) {
