+++ src/WingmanDataTrail/ListControls/MetricsSorter/EventSortByChanged.ts
