+++ src/WingmanDataTrail/ListControls/QuickSearch/EventFiltersChanged.ts
