+++ src/WingmanDataTrail/ListControls/QuickSearch/EventQuickSearchChanged.ts
